Removing weak methods
As you know, even I do not fully trust what these cryptanalysts say, The longer the big the harder to tamper with.

Of course, there maybe other methods, which involves an at least 350 years old mystery, That Fermat's last theorem , which is finally proved via a form of ecliptic curves (one of known theorems is Taniyama-Shimura theorem ).  That way may make a better matrix deterring intruders because of power.

There are  9  functions out of that 17 which are considered still not being compromised when verifying files. Of course, for password , you may need using other better methods to deter intruders.